Clearing code

+++ BisnessLayer/Interfaces/IProductable.cs
@@ -9,13 +9,8 @@ namespace BisnessLayer.Interfaces
         int? Id { get; set; }
         string Name { get; set; }
         string Category { get; set; }
-        decimal Price { get; set; }
+        float Price { get; set; }
         string Description { get; set; }
         string PictureUri { get; set; }
-
-
-        
-
-
     }
 }
